#1014 move special keywords builder operation closer to front of tests
