--- src/ch01_py/test/test_chapters_prefix_desc.py
@@ -1,11 +1,0 @@
-from src.ch01_py.chapter_desc_tools import get_chapter_desc_prefix
-
-
-def test_get_chapter_desc_prefix_ReturnsObj():
-    # ESTABLISH / WHEN / THEN
-    assert get_chapter_desc_prefix("ch03_") == "ch03"
-    assert get_chapter_desc_prefix("ch99") == "ch99"
-    assert get_chapter_desc_prefix("ch03") == "ch03"
-    assert get_chapter_desc_prefix("ch99") == "ch99"
-    assert get_chapter_desc_prefix("chXX") == "chXX"
-    assert get_chapter_desc_prefix("cha01") != "ch02"
